+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.0] - 2025-07-07
+
+### Added
+
+- ETL_engineering: `BinaryTransformer` maps string values to a binary 1 or 0 based on keyword matching.
+
+### Changed
+
+- ETL_engineering: 
+    - `KeywordDummifier` support for case insensitive regex.
+    - `RegexMapper` support for case insensitive regex.
+
 ## [3.1.0] - 2025-07-07
 
 ### Changed